add animation and correct x and y type

+++ src/ar/edu/itba/ss/Particle.java
@@ -1,27 +1,27 @@
 package ar.edu.itba.ss;
 
 public class Particle {
-    long x;
-    long y;
-    long theta;
+    double x;
+    double y;
+    double theta;
 
     int id;
 
-    public Particle(long x, long y, long theta, int id) {
+    public Particle(double x, double y, double theta, int id) {
         this.x = x;
         this.y = y;
         this.theta = theta;
         this.id = id;
     }
-    public void setTheta(long theta) {
+    public void setTheta(double theta) {
         this.theta = theta;
     }
 
-    public void setX(long x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public void setY(long y) {
+    public void setY(double y) {
         this.y = y;
     }
 }
